requirements.txt
git+https://github.com/sparkfish/augraphy
python-barcode[images]
Pillow==9.5.0
faker
torchvision